Improving performance (#117)
Improving performance

This commit tries to improve the performance of our encryption, based on the results
reported in #89.

It contains two main changes which results in improvents:

* Our pg_tde_crypt loop has several conditions to support all possible cornercases
it might encounter. These conditions are not neccessary most of the time, but slow
down our throughput considerably. This commit introduces a "simple" version of the
function for smaller data sizes, which only has a simple one instruction for loop.
* OpenSSL supports in place encryption when certain conditions are met - which is
true for our zeroblocks function. The zero block function no longer uses an
additional local array, and the filling of the input data is also simplified.

These changes together results in the following encryption overheads:

* json testcase: 140% -> 120%
* jsonb testcase: 230% -> 175%

Closes #89