Fix internal key XLogging for keyring changes (#137)
* Fix internal key XLogging for keyring changes

There is no functional relation cache during XLog replay therefore
`GetMasterKey()` would fail looking for the keyring namespace. Actually,
it would fail earlier on `get_master_key_info_path()`  as XLog replay
has its own `MyDatabaseId` and `MyDatabaseTableSpace`. The latter could
be fixed by providing `RelFileLocator` to `GetMasterKey()`  but the former
can't be fixed easily (if it could be fixed at all) as XLog replay deals
with the `Buffer` and shouldn't have access to the relations. To solve this,
we should pass the master key name in the XLog so there is no need to call
`GetMasterKey()` during XLog redo.

It also doesn't make sense to add an internal key to the cache during XLog
replay as this cache is per-process and it won't be available to any user
connections.

* Remove master_key info file on cleanup

Now tde_master_key.info is truncated which leads to the next issue:
If the file exists, `CREATE EXTENSION pg_tde;` truncates it. Then `pg_tde_set_master_key()` will create a new master key info but fail to write it to the file. Although pre-check treats an empty file as OK, the actual write opens it with `O_EXCL` (fail if file exists). Everything would work fine as the master key info is already in a cache. Until Postgres is restarted and the cache gets wiped.

This commit removes tde_master_key.info on cleanup.

* Cleanup master key info on standby