Few updates around master key versioning.
The commit contains the following changes.
Moved the logic for getting the latest available master key name
to a separate function.
Changed the base name for the master key. Now it always contains the
version number postfix.
Added a SQL interface to allow users to specify if they need to ensure
the key during set and rotate master key operations.
Also making a few adjustments in file keyring file handling.