Store rel key type along with internal key
This is needed for ALTER  TABLESPACE as pg_tdeam_relation_copy_data
gets only the old rlocator and needs to know how to properly save the
key in the new rlocator.

Also during ALTER ... TABLESPACE, tde_mdcreate gets called and
tde_smgr_get_key is trying to get the internal key. But there is no key
moved to the new table space since it is called during
RelationCreateStorage() and pg_tde.map file might not exist. And it's ok in
this case but it results in an error. This commit introduces an option not
to have a pg_tde.map (treated as no key).