Fixing warnings on the postgres side
This also modifies the pg_waldump build so it is now correctly uses
pg_tde as a static library, to handle warnings and custom flags for KMIP correctly.

This commit also enables generic -Werror for meson builds, so that we are forced
to fix warnings before merging in the future.