PG-1246: Do not duplicate makefile object targets
Issue: currently pg_tde with make wants to build the same object
files twice. Once during the pg_waldump build with Makefile.tools,
and once duriong the pg_tde build with Makefile.

As the make rulesystem sees that the object files are already built
after the first time, it doesn't rebuild them, resulting in all
kinds of issues, either incorrect pg_tde.so or a build failure.

Fix: rename the "frontend" built objects to <name>.frontend instead of
<name>.o, and duplicate the build rule for them. This way the object
files have a different name, and both will be built.

Note: the `-DFRONTEND` flag is not present in Makefile.tools, as it
is currently added by pg_waldump. The postgres side still requires
more refactoring so it feels like less of a hack, and during that,
this file can be improved further.