Fix memory usage in the internal keys cache (#385)
There was a mistake in the new cap calculations during the cache extension. It popped up only when a new cache size was multiple of a cache record (every 256 records).  Which lead to the usage of the memory beyond an allocated size. This commit fixes it along with mlock size for reallocated pages.

Also fixed a typo in a variable name.

Fixes PG-1248