CI updata 1: percona server for postgresql basic build and test
This action is similar to what we had before, executin tests on
Ubuntu 22.04 with both meson and make, all tests passing.

The test is refactored to easily support other configurations,
but those are not part of this commit yet because of failures.

Oters will be added in separate commits with required script/code
changes.

The commit also contains a compilation / possibly bug fix reported
by newer GCC versions