PG-1004 Skip encrypted files in pg_checksums
We check if a file is encrypted by looking at if it has a storage
manager encryption key for the relation file and if it has a key
we skip it.

These lookups require use to retrieve the tablespace and database
oids from the path to to the relation file.