PoC: PG-1392 Encrypt all relation forks
It is unclear to me why we stopped encrypting the FSM and visibility
map forks in commit e514ac5cc14553630b9010fe47ce77e0cdbf12c8 which
sadly does not explain why we stopped doing so. Both of them leak
metadata that we do not have to and the write load to these tables
are not heavy enough to excuse it for perfomrance reasons.

So instead of leaking data and having to jsutify that we jsut encrypt
all forks.