PG-1451: Event trigger ddl_end doesn't fire on failure
Our logic expected the start/end triggers to fire always in pairs,
always one end trigger for one start trigger.

This is not the case, and it is clearly documented in the postgres
documentation: end triggers are not executed on failed statements.

This completely broke the internal logic for encryption, and it
never worked after an error, requiring restarting the connection.

Fix: instead of relying on start-end pairs, now we check the
current transaction id both in the event trigger start function,
and during file creation. If the transaction changed during the
setup of the current event trigger data, we reset it.

There's also some additional logic around when we should and when
we can't start transactions: event triggers usually run before
statements start up transactions, and some statements, such as
DROP CONCURRENT / REINDEX CONCURRENT only work when theres no
existing transaction.