PG-1416 Throw error when no principal key for encrypted table
By first looking up the principal key and then the relation key we hid
errors when we have lost the principal key but some relations are still
encrypted. Which could for example lead to trying to read an encrypted
table as it was unencrypted causing errors like the following:

  ERROR:  invalid page in block 0 of relation base/5/16448

This does not solve the much scarier issue when we get another principal
key back from the key server but we at least get better error messages
for this common case.