PG-1447 Add tests for template databases with encrypted relations
Using a database with encrypted relations as template works works with
the WAL_LOG strateg as long as you have configured a default principal
key. It is a bit hackish but works and since encrypted objects in a
template database is likely to be an uncommon use case this is fine for
now.

The test file is put last in the test run order since there is still no
good way to clean up default principal keys.