PG-1540 Add support for CREATE and ALTER SEQUENCE to the event trigger
When creating a table with a serial/identity column the sequence will
get the same encryption status as the owning table but when doing the
same using CREATE SEQUENCE and ALTER SEQUENCE the sequence ended up not
encrypted so this add support for those.

A design choice is that sequences not owned by any table are never
encrypted which is unclear if it is a good choice or not.