Remove checking for leftover tables on extension install
This check makes no sense since tde_heap will get a new oid every time
we install it so if there are any leftovers from last time we will not
find them anyway.

Additionally for this to happen something would need to be broken with
pg_depend which I do not see why we should expend this much effort with
trying to assert against.