Update pg_buffercache_pages.c to avoid crashes
This update is to avoid crashes when the database is under high buffermapping contention. Current code does not account for when a database is under heavy usage of buffers. Recommend implementing error handling where a timeout occurs and exponential backoff reduces number of requests for the buffer header after each failure.