Make inlined hash table lookup actually make sense
There was no point in having the function for attaching the the shared
memory hash table be inline if the non-inlined code had to be called
every time anyway. We still maybe should just stop using an explicit
inline since this is unlikely to be that performace critical code but at
least if we use it we should use it correctly.