Remove license file accidentally carried over from file copy (#327)
Originally pg_tde was developed as an extension and had it's own
repository. As we progressed we moved the code from the old repo to the
contrib folder of PostgreSQL but we inadvertently left the MIT license
it was originally published under. This PR removes the MIT license and
our changes will inherit the top level license. This should avoid future
confusion about what license governs what.