Move SMGR specific logic out of the TDE map code
Especially all code realted to the keys of temporary tables did not
belong in the TDE map code and fit better in the SMGR code.

Additionally we speed up pg_tde_is_encrypted() by relying on the SMGR to
cache the relations. This might in some cases lead to blow up of the
SMGR relation cache if you query every relation in the database but
given the small size I am not overly worried.