Return void from add/change key provider functions
The returned provider id was not useful for end-users as they cannot do
anything with it. They always use the provider name when interacting
with these settings.

If they really want to see these ids they can easily just use the functions
to list all providers to see them.

The reason for this change is that it might be confusing for users when
the function to create a global provider returns -1 on success without
any indication that it's just the generated id and not an error code.