Fix pg installation reference in stormweaver run
The path had an additional "../" section, which means it referenced
outside the action working directory. For some reason (possibly earlier
scripts during development?) we had an existing postgres installation
there, but it was old and never updated. As it still had  bugs in the
partition handling code, it always reported an error.