Refactor WAL SMGR to make it easier to add WAL write support to fronend tools
With ifdefs all over the place it was hard to expose the write functions
to frontend tools so we reduce the number of ifdefs by having one clear
set of data structures fror backend and one for frontend.

Additionally we give access to WAL key generation and setting the start_lsn
of a key to the frontend code.