PG-1710 Decrypt partial WAL segments when archvining them
While the restore_command will never be called for partial WAL segments
we should sitll make sure to decrypt them when archiving so a sysadmin
manually could use them. As the comemnt explains we also add the same
logic to the restore command for symmetry and if someone ever would call
it manually or PostgreSQL would add support for restoring partial WAL
segments.

If you want to learn more about partial WAL segments read the comment
for CleanupAfterArchiveRecovery().